Fix no checkout if dir exists
When the destination directory already existed the script would checkout
anyway and wipe all files in it. To avoid this problem the script only
checks out if the directory does not exist.

To correctly fix this issue the script should check if the directory
exists and if the working copy URL is the same as the input. For now
this fix should suffice.